Jury: Modul G

+++ src/main/java/ch/zkmf2024/server/dto/JudgeReportScoreDTO.java
@@ -1,11 +1,12 @@
 package ch.zkmf2024.server.dto;
 
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public record JudgeReportScoreDTO(@NotNull Long reportId,
                                   @NotNull String judgeName,
                                   @NotNull String judgeRole,
-                                  Integer score,
+                                  BigDecimal score,
                                   boolean ratingFixed,
                                   boolean done) {
 }
